+++ packages/create-fleur-next-app/template/lib/fleur.tsx
@@ -1,8 +1,13 @@
-import { default as NextApp } from 'next/app'
+import NextApp, { AppContext, AppProps, AppInitialProps } from 'next/app'
 import { FleurContext } from '@fleur/react'
 import { useMemo } from 'react'
-import { bindFleurContext } from '@fleur/next'
-import { createContext } from '../domains/app'
+import {
+  bindFleurContext,
+  serializeContext,
+  deserializeContext,
+  PageContext,
+} from '@fleur/next'
+import { createContext } from '../domains'
 
 export const getOrCreateFleurContext = (state: any = null) => {
   const isServer = typeof window === 'undefined'
@@ -15,10 +20,21 @@ export const getOrCreateFleurContext = (state: any = null) => {
   return context
 }
 
-export const appWithFleurContext = (App: typeof NextApp) => {
+export type FleurAppContext = AppContext & { ctx: PageContext }
+
+declare class ClassApp extends NextApp {
+  static getInitialProps(appContext: FleurAppContext): Promise<AppInitialProps>
+}
+
+interface FunctionApp {
+  (props: AppProps): JSX.Element
+  getInitialProps(appContext: FleurAppContext): Promise<AppInitialProps>
+}
+
+export const appWithFleurContext = (App: typeof ClassApp | FunctionApp) => {
   const Comp = ({ __FLEUR_STATE__, ...props }: any) => {
     const fleurContext = useMemo(
-      () => getOrCreateFleurContext(__FLEUR_STATE__),
+      () => getOrCreateFleurContext(deserializeContext(__FLEUR_STATE__)),
       [],
     )
 
@@ -37,7 +53,7 @@ export const appWithFleurContext = (App: typeof NextApp) => {
 
     return {
       ...appProps,
-      __FLEUR_STATE__: fleurContext.dehydrate(),
+      __FLEUR_STATE__: serializeContext(fleurContext),
     }
   }